@@ -1,97 +1,79 @@
 source 'https://rubygems.org'
-ruby '2.3.0'
 
-# Bundle edge Rails instead: gem 'rails', github: 'rails/rails'
-gem 'rails', '4.2.0'
+ruby '2.4.2'
+gem 'rails', '~> 5.1'
 
-group :production do
-	# MySQL
-	gem 'mysql2', '~> 0.3.18'
-end
-
-gem 'rails_12factor', group: :production
-
-# Use SCSS for stylesheets
-gem 'sass-rails', '~> 4.0.3'
-
-# Use Uglifier as compressor for JavaScript assets
+# Required gems for rails
+gem 'jbuilder', '~> 2.7'
 gem 'uglifier', '>= 1.3.0'
+gem 'sass-rails', '>= 5.0.5'
+gem 'puma', '~> 3.10'
+gem 'coffee-rails', '~> 4.2'
+# gem "turbolinks", "~> 5"
+gem 'tzinfo-data', platforms: [:mswin, :mingw, :x64_mingw]
 
-# Use jQuery as the JavaScript library
-gem 'jquery-rails'
-
-# Add jQuery UI gem
-gem 'jquery-ui-rails'
-
-# Build JSON APIs with ease. Read more: https://github.com/rails/jbuilder
-gem 'jbuilder', '~> 2.0'
-# bundle exec rake doc:rails generates the API under doc/api.
-
-gem 'sdoc', '~> 0.4.0',          group: :doc
-
-# Spring speeds up development by keeping your application running in the background. Read more: https://github.com/rails/spring
-# gem 'spring',        group: :development
-
-gem 'puma'
-
-# explicitly include the cofee-script gem to prevent implicit require that result in the following message:
-# "WARN: tilt autoloading 'coffee_script' in a non thread-safe way; explicit require 'coffee_script' suggested."
-gem 'coffee-rails'
-
-group :development, :test do
-  # Windows developers, refer to: https://github.com/sparklemotion/sqlite3-ruby/issues/185 to deal with
-  # `require': cannot load such file -- sqlite3/sqlite3_native (LoadError)
-  # Essentially, you just need to build sqlite3 >= 1.3.11 and uninstall the bundled mingw sqlite.
-  gem 'sqlite3'
-end
+# Note:
+# The mingw version of bcrypt does not work correctly on Windows.
+# A temporary workaround is to run the following command:
+# gem uninstall bcrypt --all --force && gem install bcrypt --platform=ruby
 
 #####################################
 ####### Actual Carpe Gems ###########
 #####################################
 
+# bundle exec rails doc:rails generates the API under doc/api.
+gem 'sdoc', group: :doc
+
+# Databases
+gem 'mysql2', group: :production
+gem 'sqlite3', group: [:development, :test]
+
 # Use devise for authentication
-gem 'devise'
+gem 'devise', '~> 4.3'
 
 # Use omniauth to allow different logins
-gem 'omniauth'
-
-# Google omniauth
+gem 'omniauth', '~> 1.7'
 gem 'omniauth-google-oauth2'
 
 # Use local-time for timezone handling, most likely unused ATM
-gem 'local_time'
+gem 'local_time', '~> 2.0'
 
 # Use rubycritic to detect code smell and problems (https://github.com/whitesmith/rubycritic)
 gem 'rubycritic', :require => false, group: :development
 
 # Use chartkick to make graphs using the Google Graph API
-gem 'chartkick'
+gem 'chartkick', '~> 2.2'
 
 # Roadie, used for styling emails nicely
-gem 'roadie', '~> 3.1.1'
+gem 'roadie', '~> 3.2'
 
 # Use rack-mini-profiler for investigating site speed (https://github.com/MiniProfiler/rack-mini-profiler)
 gem 'rack-mini-profiler', group: [:development, :production]
 
 # Paperclip gem for managing user uploaded images and such (https://github.com/thoughtbot/paperclip)
-gem 'paperclip'
+gem 'paperclip', '~> 5.1'
 
 # AWS SDK gem for connecting to Amazon S3 and other tools (https://github.com/aws/aws-sdk-ruby)
-gem 'aws-sdk', '~> 2', group: :production
-
-if Gem.win_platform?
-  # Used by Windows for time zone differences (strange indeed).
-  gem 'tzinfo-data'
-
-  gem 'coffee-script-source', '>= 1.12.2'
-end
+gem 'aws-sdk', '~> 3', group: :production
 
 # Add SimpleCov to check test coverage (https://github.com/colszowka/simplecov)
 gem 'simplecov', :require => false, :group => :test
 
 # Add teaspoon gem for JS testing (https://github.com/jejacks0n/teaspoon) with Mocha
-gem "teaspoon"
+gem "teaspoon", '~> 1.1'
 gem "teaspoon-mocha"
 
 # Minitest reporting
 gem "minitest-reporters"
+
+# LSG
+gem "livingstyleguide"
+
+# A more informative exception template page
+gem 'better_errors', group: :development
+
+# Use jQuery as the JavaScript library
+gem 'jquery-rails', '~> 4.3'
+
+# Add jQuery UI gem
+gem 'jquery-ui-rails', '~> 6.0'
