@@ -1,26 +1,7 @@
+require 'tasky'
+
 class PagesController < ApplicationController
-  before_filter :authorize_admin,
-                only: [:promote, :admin, :sandbox, :destroy_user, :admin_user_info]
-
-  def promote # promote or demote users admin status
-    @user = User.find(params[:id])
-    @user.admin = (params[:de] != "true") # they're an admin if not being demoted
-    @user.save
-
-    render json: { action: "promote", uid: params[:id] }
-  end
-
-  def destroy_user
-    user = User.from_param params[:id]
-    user.destroy
-
-    redirect_to "/admin"
-  end
-
-  def admin_user_info
-    @user = User.from_param params[:id]
-    @user_groups = UsersGroup.where(user_id: @user.id)
-  end
+  before_action :authorize_admin!, only: [:admin, :sandbox]
 
   def admin #admin page
     @now = Time.zone.now
@@ -31,32 +12,28 @@ class PagesController < ApplicationController
     @past_month_events_modified = Event.where('created_at >= ?', Time.zone.now - 1.months).group(:updated_at).count
   end
 
-  # runs a command based on a given admin panel button id
-  # button id is passed via params[:button_id]
+  # Runs predefined server commands requested from the admin panel.
   def run_command
-    cmd =
-      case params[:button_id]
-      when "run-jsdoc"               then "npm run jsdoc"
-      when "run-js-unit-tests"       then "npm run teaspoon"
-      when "run-js-acceptance-tests" then "npm run codeceptjs"
-      when "run-rails-unit-tests"    then "bundle exec rake test"
-      end
+    cmd = case params[:button_id]
+          when "run-jsdoc"               then "npm run jsdoc --silent"
+          when "run-js-unit-tests"       then "npm run teaspoon --silent"
+          when "run-js-acceptance-tests" then "npm run acceptance-tests --silent"
+          when "run-rails-unit-tests"    then "npm run minitest --silent"
+          end
 
-    pid = Process.spawn cmd
-    Process.detach pid # prevents zombie processes
-    render json: { button_id: params[:button_id], pid: pid }
-  rescue
-    render json: { button_id: params[:button_id], pid: pid, error: "true" }
+    task_id = Tasky::run cmd
+    render json: params.merge(task_id: task_id)
+  rescue Tasky::CommandError => e
+    render json: params.merge(cmd_error: e.inspect)
   end
 
-  # Checks if a command run on a given pid is complete
-  # pid is grabbed from params[:pid]
   def check_if_command_is_finished
-    # signal 0 checks if the processor exists
-    Process.kill 0, params[:pid].to_i
-    render json: { finished: "false" }
-  rescue
-    # (mostly likely) the process no longer exists
-    render json: { finished: "true" }
+    task = Tasky::fetch_task params[:task_id]
+
+    if task.finished?
+      render json: {log: (task.success? ? "SUCCESS" : task.error_log)}
+    else
+      render json: {check_again: true}
+    end
   end
 end
